Finish 6.2.0 compiler packaging (#13)
* Add package for experimental header files.

Unified runtime depends on experimental header files from
compute runtime level-zero. This package delivers those
headers so we do not have to vendor this dependency.

* Use exp headers from PPA

* Bump compiler version

* Disable LLVM_BUILD_LLVM_DYLIB

* Bump LLVM/Clang to 21

* Package source code for vc-intrinsics.

The new version of the DPC++ compiler is based on LLVM/Clang v21,
while the latest version available in the archive is v20. Creating
this package will allow vc-intrinics to be built in-tree with
the compiler and ensure better compatability.

* Fix install path

* Build compiler with vc-intrinsics in-tree

* intel-vc-intrinsics-src: include CMakeLists.txt

* Fix typo in install path

* Fix install and link filenames, bump compiler version

* Remove scripts for managing vendored deps

* Transition compiler packages from previous release

* Add missing runtime dependencies for compiler

* Update README for 6.2.0 release

* Clean up compiler build tooling

* Vendor vc-intrinsics and level-zero headers

* Add more detail about L0 vendoring